+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/MainScaffold.kt
@@ -1,0 +1,20 @@
+package com.whatever.caramel.feat.main
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+internal fun MainScaffold(
+    modifier: Modifier = Modifier,
+    bottomBar: @Composable () -> Unit,
+    content: @Composable (PaddingValues) -> Unit
+) {
+    Scaffold(
+        modifier = modifier,
+        bottomBar = bottomBar
+    ) { padding ->
+        content(padding)
+    }
+}
