+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/calendar/mvi/CalendarState.kt
@@ -1,0 +1,7 @@
+package com.whatever.caramel.feat.main.calendar.mvi
+
+import com.whatever.caramel.core.presentation.UiState
+
+data class CalendarState(
+    val text: String = ""
+) : UiState
