+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/home/mvi/HomeState.kt
@@ -1,0 +1,7 @@
+package com.whatever.caramel.feat.main.home.mvi
+
+import com.whatever.caramel.core.presentation.UiState
+
+data class HomeState(
+    val text: String = ""
+) : UiState
