+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/memo/MemoScreen.kt
@@ -1,0 +1,43 @@
+package com.whatever.caramel.feat.main.memo
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.sp
+import com.whatever.caramel.feat.main.memo.mvi.MemoIntent
+import com.whatever.caramel.feat.main.memo.mvi.MemoState
+
+@Composable
+internal fun MemoScreen(
+    state: MemoState,
+    onIntent: (MemoIntent) -> Unit
+) {
+    Box(
+        modifier = Modifier.fillMaxSize()
+    ) {
+        Text(
+            modifier = Modifier.align(alignment = Alignment.Center),
+            text = "메모 화면 입니다.",
+            fontSize = 32.sp
+        )
+
+        Button(
+            modifier = Modifier
+                .fillMaxWidth()
+                .align(
+                    alignment = Alignment.BottomCenter
+                ),
+            onClick = { onIntent(MemoIntent.ClickMemo) }
+        ) {
+            Text(
+                text = "메모 아이템",
+                fontSize = 12.sp
+            )
+        }
+    }
+}
