+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/memo/mvi/MemoIntent.kt
@@ -1,0 +1,9 @@
+package com.whatever.caramel.feat.main.memo.mvi
+
+import com.whatever.caramel.core.presentation.UiIntent
+
+sealed interface MemoIntent : UiIntent {
+
+    data object ClickMemo : MemoIntent
+
+}
