+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/memo/mvi/MemoSideEffect.kt
@@ -1,0 +1,9 @@
+package com.whatever.caramel.feat.main.memo.mvi
+
+import com.whatever.caramel.core.presentation.UiSideEffect
+
+sealed interface MemoSideEffect : UiSideEffect {
+
+    data object NavigateToTodoDetail : MemoSideEffect
+
+}
