+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/memo/mvi/MemoState.kt
@@ -1,0 +1,7 @@
+package com.whatever.caramel.feat.main.memo.mvi
+
+import com.whatever.caramel.core.presentation.UiState
+
+data class MemoState(
+    val text: String = ""
+) : UiState
