--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/MainScaffold.kt
@@ -1,20 +1,0 @@
-package com.whatever.caramel.feat.main.presentation
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-internal fun MainScaffold(
-    modifier: Modifier = Modifier,
-    bottomBar: @Composable () -> Unit,
-    content: @Composable (PaddingValues) -> Unit
-) {
-    Scaffold(
-        modifier = modifier,
-        bottomBar = bottomBar
-    ) { padding ->
-        content(padding)
-    }
-}
