--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/calendar/CalendarViewModel.kt
@@ -1,24 +1,0 @@
-package com.whatever.caramel.feat.main.presentation.calendar
-
-import androidx.lifecycle.SavedStateHandle
-import com.whatever.caramel.core.presentation.BaseViewModel
-import com.whatever.caramel.feat.main.presentation.calendar.mvi.CalendarIntent
-import com.whatever.caramel.feat.main.presentation.calendar.mvi.CalendarSideEffect
-import com.whatever.caramel.feat.main.presentation.calendar.mvi.CalendarState
-
-class CalendarViewModel(
-    savedStateHandle: SavedStateHandle
-) : BaseViewModel<CalendarState, CalendarSideEffect, CalendarIntent>(savedStateHandle) {
-
-    override fun createInitialState(savedStateHandle: SavedStateHandle): CalendarState {
-        return CalendarState()
-    }
-
-    override suspend fun handleIntent(intent: CalendarIntent) {
-        when (intent) {
-            is CalendarIntent.ClickTodoCard -> postSideEffect(CalendarSideEffect.NavigateToCreateTodo)
-            is CalendarIntent.ClickCreateTodoButton -> postSideEffect(CalendarSideEffect.NavigateToTodoDetail)
-        }
-    }
-
-}
