--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/calendar/mvi/CalendarState.kt
@@ -1,7 +1,0 @@
-package com.whatever.caramel.feat.main.presentation.calendar.mvi
-
-import com.whatever.caramel.core.presentation.UiState
-
-data class CalendarState(
-    val text: String = ""
-) : UiState
