--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/home/mvi/HomeIntent.kt
@@ -1,15 +1,0 @@
-package com.whatever.caramel.feat.main.presentation.home.mvi
-
-import com.whatever.caramel.core.presentation.UiIntent
-
-sealed interface HomeIntent : UiIntent {
-
-    data object ClickSettingButton : HomeIntent
-
-    data object ClickStartedCoupleDayButton : HomeIntent
-
-    data object ClickCreateTodoItem : HomeIntent
-
-    data object ClickTodoItem : HomeIntent
-
-}
