--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/home/mvi/HomeSideEffect.kt
@@ -1,15 +1,0 @@
-package com.whatever.caramel.feat.main.presentation.home.mvi
-
-import com.whatever.caramel.core.presentation.UiSideEffect
-
-sealed interface HomeSideEffect : UiSideEffect {
-
-    data object NavigateToSetting : HomeSideEffect
-
-    data object NavigateToCreateTodo : HomeSideEffect
-
-    data object NavigateToTodoDetail : HomeSideEffect
-
-    data object NavigateToStartedCoupleDay : HomeSideEffect
-
-}
