--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/home/mvi/HomeState.kt
@@ -1,7 +1,0 @@
-package com.whatever.caramel.feat.main.presentation.home.mvi
-
-import com.whatever.caramel.core.presentation.UiState
-
-data class HomeState(
-    val text: String = ""
-) : UiState
