--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/memo/mvi/MemoIntent.kt
@@ -1,9 +1,0 @@
-package com.whatever.caramel.feat.main.presentation.memo.mvi
-
-import com.whatever.caramel.core.presentation.UiIntent
-
-sealed interface MemoIntent : UiIntent {
-
-    data object ClickMemo : MemoIntent
-
-}
