--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/presentation/memo/mvi/MemoState.kt
@@ -1,7 +1,0 @@
-package com.whatever.caramel.feat.main.presentation.memo.mvi
-
-import com.whatever.caramel.core.presentation.UiState
-
-data class MemoState(
-    val text: String = ""
-) : UiState
